+++ __tests__/cairo1.test.ts
@@ -5,6 +5,7 @@ import {
   compiledHelloSierra,
   compiledHelloSierraCasm,
   describeIfDevnetSequencer,
+  describeIfSequencerTestnet2,
   getTestAccount,
   getTestProvider,
 } from './fixtures';
@@ -12,11 +13,11 @@ import { initializeMatcher } from './schema';
 
 // Testnet will not accept declare v2 with same compiledClassHash,
 // aka. we can't redeclare same contract
-describeIfDevnetSequencer('Cairo 1', () => {
+describeIfDevnetSequencer('Cairo 1 Devnet', () => {
   describe('Sequencer API', () => {
     const provider = getTestProvider() as SequencerProvider;
     const account = getTestAccount(provider);
-    let classHash: any; // = '0x5b3507904c785fcceff17b34b4269f729bbddc1a432e4a63145c70071383413';
+    let classHash: any; // = '0x3e2e625998f89befe4d429d5d958275f86421310bfb00440c2431140e8c90ba';
     let contractAddress: any;
     let declareV2Tx: any;
     let cairo1Contract: Contract;
@@ -83,14 +84,28 @@ describeIfDevnetSequencer('Cairo 1', () => {
       const tx = await cairo1Contract.increase_balance(100);
       await account.waitForTransaction(tx.transaction_hash);
       const balance = await cairo1Contract.get_balance();
-      expect(toBigInt(balance)).toBe(200n);
+      expect(balance).toBe(200n);
     });
 
-    test('Cairo 1 Contract Interaction - uint', async () => {
-      const tx = await cairo1Contract.increase_balance_u8(255);
+    test('Cairo 1 Contract Interaction - uint 8, 16, 32, 64, 128', async () => {
+      const tx = await cairo1Contract.increase_balance_u8(255n);
       await account.waitForTransaction(tx.transaction_hash);
       const balance = await cairo1Contract.get_balance_u8();
-      expect(toBigInt(balance)).toBe(255n);
+      expect(balance).toBe(255n);
+
+      let result = await cairo1Contract.test_u16(255n);
+      expect(result).toBe(256n);
+      result = await cairo1Contract.test_u32(255n);
+      expect(result).toBe(256n);
+      result = await cairo1Contract.test_u64(255n);
+      expect(result).toBe(256n);
+      result = await cairo1Contract.test_u128(255n);
+      expect(result).toBe(256n);
+    });
+
+    test('Cairo 1 - uint256 struct', async () => {
+      const result = await cairo1Contract.test_u256(2n ** 256n - 2n);
+      expect(result).toBe(2n ** 256n - 1n);
     });
 
     test('Cairo 1 Contract Interaction - bool', async () => {
@@ -122,7 +137,6 @@ describeIfDevnetSequencer('Cairo 1', () => {
     });
 
     test('Cairo 1 Contract Interaction - echo flat un-named un-nested tuple', async () => {
-      // TODO: flatten result ?
       const status = await cairo1Contract.echo_un_tuple(tuple(77, 123));
       expect(Object.values(status)).toEqual([77n, 123n]);
     });
@@ -139,6 +153,66 @@ describeIfDevnetSequencer('Cairo 1', () => {
         val: 'simple',
       });
       expect(status).toBe('simple');
+    });
+  });
+});
+
+describeIfSequencerTestnet2('Cairo1 Testnet2', () => {
+  describe('Sequencer API', () => {
+    const provider = getTestProvider() as SequencerProvider;
+    const account = getTestAccount(provider);
+    const classHash: any = '0x028b6f2ee9ae00d55a32072d939a55a6eb522974a283880f3c73a64c2f9fd6d6';
+    const contractAddress: any =
+      '0x771bbe2ba64fa5ab52f0c142b4296fc67460a3a2372b4cdce752c620e3e8194';
+    let cairo1Contract: Contract;
+    initializeMatcher(expect);
+
+    beforeAll(async () => {
+      const cairoClass = await provider.getClassByHash(classHash);
+      cairo1Contract = new Contract(cairoClass.abi, contractAddress, account);
+    });
+
+    test('getCompiledClassByClassHash', async () => {
+      const compiledClass = await provider.getCompiledClassByClassHash(classHash);
+      expect(compiledClass).toMatchSchemaRef('CompiledClass');
+    });
+
+    test('GetClassByHash', async () => {
+      const classResponse = await provider.getClassByHash(classHash);
+      expect(classResponse).toMatchSchemaRef('SierraContractClass');
+    });
+
+    test('GetClassAt', async () => {
+      const classResponse = await provider.getClassAt(contractAddress);
+      expect(classResponse).toMatchSchemaRef('SierraContractClass');
+    });
+
+    test('Cairo 1 Contract Interaction - felt252', async () => {
+      const result = await cairo1Contract.test_felt252(100);
+      expect(result).toBe(101n);
+    });
+
+    test('Cairo 1 Contract Interaction - uint 8, 16, 32, 64, 128', async () => {
+      let result = await cairo1Contract.test_u8(100n);
+      expect(result).toBe(107n);
+      result = await cairo1Contract.test_u16(100n);
+      expect(result).toBe(106n);
+      result = await cairo1Contract.test_u32(100n);
+      expect(result).toBe(104n);
+      result = await cairo1Contract.test_u64(255n);
+      expect(result).toBe(258n);
+      result = await cairo1Contract.test_u128(255n);
+      expect(result).toBe(257n);
+    });
+
+    test('Cairo 1 - uint256 struct', async () => {
+      const result = await cairo1Contract.test_u256(2n ** 256n - 2n);
+      expect(result).toBe(2n ** 256n - 1n);
+    });
+
+    test('Cairo 1 Contract Interaction - bool', async () => {
+      const tx = await cairo1Contract.test_bool();
+      expect(tx).toBe(true);
     });
   });
 });