+++ src/utils/calldata/responseParser.ts
@@ -1,6 +1,8 @@
+/* eslint-disable no-case-declarations */
 import { AbiEntry, AbiStructs, Args, ParsedStruct } from '../../types';
 import { BigNumberish } from '../num';
-import { isCairo1Type, isLen, isTypeArray, isTypeBool, isTypeTuple } from './cairo';
+import { uint256ToBN } from '../uint256';
+import { isCairo1Type, isLen, isTypeArray, isTypeBool, isTypeTuple, isTypeUint256 } from './cairo';
 import extractTupleMemberTypes from './tuple';
 
 /**
@@ -62,6 +64,10 @@ export default function responseParser(
     case isTypeBool(type):
       temp = responseIterator.next().value;
       return Boolean(BigInt(temp));
+    case isTypeUint256(type):
+      const low = responseIterator.next().value;
+      const high = responseIterator.next().value;
+      return uint256ToBN({ low, high });
     case isTypeArray(type):
       // eslint-disable-next-line no-case-declarations
       const parsedDataArr: (BigNumberish | ParsedStruct)[] = [];